# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RomaniaDate Dating App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /matches/{matchId}:
    delete:
      operationId: unmatch
      tags: [matches]
      summary: Unmatch
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unmatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /chat/{matchId}/messages:
    get:
      operationId: getMessages
      tags: [chat]
      summary: Get messages for a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        birthDate:
          type: string
        gender:
          type: string
          enum: [boy, girl]
        preference:
          type: string
          enum: [boy, girl, both]
        city:
          type: string
        referralCode:
          type: ["string", "null"]
      required:
        - email
        - password
        - name
        - birthDate
        - gender
        - preference
        - city
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        birthDate:
          type: string
        age:
          type: integer
        ageCategory:
          type: string
          enum: [teen, adult]
        gender:
          type: string
          enum: [boy, girl]
        preference:
          type: string
          enum: [boy, girl, both]
        city:
          type: string
        bio:
          type: ["string", "null"]
        photos:
          type: array
          items:
            type: string
        interests:
          type: array
          items:
            type: string
        diamonds:
          type: integer
        isVip:
          type: boolean
        vipExpiresAt:
          type: ["string", "null"]
        isVerified:
          type: boolean
        isPrivate:
          type: boolean
        isBoosted:
          type: boolean
        boostExpiresAt:
          type: ["string", "null"]
        level:
          type: integer
        xp:
          type: integer
        rank:
          type: ["string", "null"]
        referralCode:
          type: string
        isOnline:
          type: boolean
        lastSeen:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        likesGivenToday:
          type: integer
        superlikesGivenToday:
          type: integer
        notifyNewMatch:
          type: boolean
        notifyNewMessage:
          type: boolean
        notifyDiamonds:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - birthDate
        - age
        - ageCategory
        - gender
        - preference
        - city
        - photos
        - interests
        - diamonds
        - isVip
        - isVerified
        - isPrivate
        - isBoosted
        - level
        - xp
        - referralCode
        - isOnline
        - isAdmin
        - likesGivenToday
        - superlikesGivenToday
        - notifyNewMatch
        - notifyNewMessage
        - notifyDiamonds
        - createdAt
...
    Message:
      type: object
      properties:
        id:
          type: integer
        matchId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        type:
          type: string
          enum: [text, image, emoji]
        isRead:
          type: boolean
        isDelivered:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - matchId
        - senderId
        - content
        - type
        - isRead
        - isDelivered
        - createdAt
    SendMessageRequest:
      type: object
      properties:
        content:
          type: string
        type:
          type: string
          enum: [text, image, emoji]
      required:
        - content
        - type
...