# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RomaniaDate Dating App API
 * OpenAPI spec version: 0.1.0
 */
import type { UserAgeCategory } from "./userAgeCategory";
import type { UserGender } from "./userGender";
import type { UserPreference } from "./userPreference";

export interface User {
  id: number;
  email: string;
  name: string;
  birthDate: string;
  age: number;
  ageCategory: UserAgeCategory;
  gender: UserGender;
  preference: UserPreference;
  city: string;
  /** @nullable */
  bio?: string | null;
  photos: string[];
  interests: string[];
  diamonds: number;
  isVip: boolean;
  /** @nullable */
  vipExpiresAt?: string | null;
  isVerified: boolean;
  isPrivate: boolean;
  isBoosted: boolean;
  /** @nullable */
  boostExpiresAt?: string | null;
  level: number;
  xp: number;
  /** @nullable */
  rank?: string | null;
  referralCode: string;
  isOnline: boolean;
  /** @nullable */
  lastSeen?: string | null;
  isAdmin: boolean;
  likesGivenToday: number;
  superlikesGivenToday: number;
  notifyNewMatch: boolean;
  notifyNewMessage: boolean;
  notifyDiamonds: boolean;
  createdAt: string;
}